Fix NSUInteger / unsigned mismatch.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/apps/gorm/trunk@36031 72102866-910b-0410-8b05-ffd578937521

+++ GormCore/GormClassEditor.m
@@ -465,7 +465,7 @@ NSImage *browserImage = nil;
 }
 
 //--- IBSelectionOwners protocol ---
-- (unsigned) selectionCount
+- (NSUInteger) selectionCount
 {
   return ([outlineView selectedRow] == -1)?0:1;
 }
